HHH-10873 - Type systems consolidation + org.hibernate.type refactoring HHH-11478 - General package restructuring HHH-11263 - Improvements in boot-time model and its creation as well as processing into runtime-model
